Graph zooming and dragging.
Now it is possible to zoom in and out using the tool bar buttons, also,
I added the functionality to drag the graph and move through it. I
changed the x axis back to date type because the labels were wrong, but
now it has the multiple lane error again.